pyproject.toml
[tool.poetry]
name = "moderate_api"
version = "0.2.0"
description = "HTTP API of the MODERATE platform"
authors = ["Andres Garcia Mangas <andres.garcia@fundacionctic.org>"]
license = "EUPL-1.2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.8"
uvicorn = { extras = ["standard"], version = "^0.34.0" }
coloredlogs = "^15.0.1"
pyjwt = "^2.8.0"
httpx = "^0.27.0"
cryptography = "^41.0.5"
cachetools = "^5.3.2"
asyncache = "^0.3.1"
pydantic = "^1.10.0"
python-multipart = "^0.0.6"
itsdangerous = "^2.1.2"
ujson = "^5.8.0"
orjson = "^3.9.10"
email-validator = "^2.1.0.post1"
asyncpg = "^0.29.0"
sqlmodel = "^0.0.21"
greenlet = "^3.0.1"
casbin = "^1.33.0"
arrow = "^1.3.0"
aiobotocore = "^2.7.0"
python-slugify = "^8.0.1"
pygwalker = "^0.4.6"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
numpy = "^1.26.4"
marimo = "^0.13.0"
polars = "^1.3.0"
aio-pika = "^9.4.3"
matplotlib = "^3.8.0"
pandas = "^2.1.1"
zstandard = "^0.23.0"
seaborn = "^0.13.2"
torch = [
    { version = "^2.6.0", platform = "darwin", source = "pypi" },
    { version = "^2.6.0", platform = "win32", source = "pypi" },
    { version = "^2.6.0", platform = "linux", source = "pytorch-cpu" },
]
plotly = "^6.1.0"
numba = "^0.61.2"
scipy = "^1.15.3"
scikit-learn = "^1.7.0"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
pytest = "^8.3.0"
pytest-asyncio = "^0.23.0"
pyarrow = "^14.0.0"
ruff = "^0.4.4"
mypy = "~1.8.0"
types-requests = "^2.32.4.20250913"
types-cachetools = "^6.2.0.20251022"

[tool.ruff]
line-length = 79
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
exclude = ["moderate_api/notebooks/"]

# Relax some checks for SQLAlchemy-heavy files with complex generic types
[[tool.mypy.overrides]]
module = [
    "moderate_api.entities.crud",
    "moderate_api.entities.asset.search",
    "moderate_api.entities.asset.router",
    "moderate_api.entities.asset.models",
]
disallow_any_generics = false
disallow_untyped_defs = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
write-openapi = "moderate_api.openapi:write_openapi"
